+++ src/Filter/StringFilter/StringFilterInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+  declare(strict_types=1);
+
+  namespace Fiv\Form\Filter\StringFilter;
+
+  interface StringFilterInterface {
+
+    public function apply(string $input) : string;
+
+  }