+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 ## 0.1.0-alpha.8 [Unreleased]
-
+### Added
+  - #49 Add new filter interface `Fiv\Form\Filter\StringFilter\StringFilterInterface`
+    
 ## 0.1.0-alpha.7 [2017-02-01]
 ### Added
   - Add new element class `\Fiv\Form\Element\Hidden`